fancy installer! fixes #3; fixes #2; fixes #4; fixes #5; fixes #6

+++ install.sh
@@ -14,14 +14,14 @@
 #        AUTHOR: Pascal Hebbeker (PH), pascal.hebbeker@gmail.com
 #  ORGANIZATION: 
 #       CREATED: 2016-09-15 09:39
-#      REVISION: 2016-09-15 11:30
+#      REVISION: 2016-09-15 14:57
 #===============================================================================
 
 set -o nounset                              # Treat unset variables as an error
 set -e
 
 e_header()   { echo -e "\e[35m\n$@\e[0m"; }
-e_success()  { echo -e " \e[32m✓\e[0m $@"; }
+e_success()  { echo -e " \e[32m*\e[0m $@"; }
 e_error()    { echo -e " \e[31mx\e[0m $@"; exit 1 ; }
 e_warn()    { echo -e " \e[31m!\e[0m $@"; }
 e_arrow()    { echo -e " \e[33m>\e[0m $@"; }
@@ -49,53 +49,181 @@ git pull $REPOSRC
 fi
 cd $curdir
 }
+cpfile () {
+   cp $name/$dotfile ~/.$dotfile
+}
 
-#progs=(vim tmux shell sol gitlab)
-progs=(pvim ptmux psol)
+progs=(pvim ptmux pshell psol pgitlab)
 
 pvim() {
 name="vim"
-if [ "$@" = "dotfile" ] ; then
-   dotfile="vimrc"
-   dotdir="vim"
-fi
+dotfile="vimrc"
+dotdir="vim"
 
 if [ "$@" = "install" ] ; then
-   e_arrow2 "cloning vundle"
-   gitclone https://github.com/vundlevim/vundle.vim.git ~/.vim/bundle/vundle.vim
-   e_arrow2 "installing plugins"
-   vim +"set viewoptions=" +plugininstall +qall
-   e_arrow2 "updating template files accoring to $conffile"
-   tfiles=( ~/.vim/bundle/bash-support.vim/bash-support/templates/templates ~/.vim/bundle/awk-support.vim/awk-support/templates/templates )
-
+   cpfile
+   e_arrow2 "Cloning Vundle"
+   gitclone https://github.com/VundleVim/Vundle.vim.git ~/.vim/bundle/Vundle.vim
+   e_arrow2 "Installing Plugins"
+   vim +"set viewoptions=" +PluginInstall +qall
+   e_arrow2 "Updating Template Files according to $CONFFILE"
+   tfiles=( ~/.vim/bundle/bash-support.vim/bash-support/templates/Templates ~/.vim/bundle/awk-support.vim/awk-support/templates/Templates )
    for f in ${tfiles[@]}; do 
-      sed -i "s/^\(setmacro( 'author', *'\).*\('\)/\1$tauthor\2/" $f
-      sed -i "s/^\(setmacro( 'authorref', *'\).*\('\)/\1$tauthorref\2/" $f
-      sed -i "s/^\(setmacro( 'company', *'\).*\('\)/\1$tcompany\2/" $f
-      sed -i "s/^\(setmacro( 'copyright', *'\).*\('\)/\1$tcopyright\2/" $f
-      sed -i "s/^\(setmacro( 'email', *'\).*\('\)/\1$temail\2/" $f
-      sed -i "s/^\(setmacro( 'license', *'\).*\('\)/\1$tlicense\2/" $f
-      sed -i "s/^\(setmacro( 'organization', *'\).*\('\)/\1$torganization\2/" $f
-      sed -i "s/^\(setformat( 'date', *'\).*\('\)/\1$tdate\2/" $f
-      sed -i "s/^\(setformat( 'time', *'\).*\('\)/\1$ttime\2/" $f
-      sed -i "s/^\(setformat( 'year', *'\).*\('\)/\1$tyear\2/" $f
+      sed -i "s/^\(SetMacro( 'AUTHOR', *'\).*\('\)/\1$tauthor\2/" $f
+      sed -i "s/^\(SetMacro( 'AUTHORREF', *'\).*\('\)/\1$tauthorref\2/" $f
+      sed -i "s/^\(SetMacro( 'COMPANY', *'\).*\('\)/\1$tcompany\2/" $f
+      sed -i "s/^\(SetMacro( 'COPYRIGHT', *'\).*\('\)/\1$tcopyright\2/" $f
+      sed -i "s/^\(SetMacro( 'EMAIL', *'\).*\('\)/\1$temail\2/" $f
+      sed -i "s/^\(SetMacro( 'LICENSE', *'\).*\('\)/\1$tlicense\2/" $f
+      sed -i "s/^\(SetMacro( 'ORGANIZATION', *'\).*\('\)/\1$torganization\2/" $f
+      sed -i "s/^\(SetFormat( 'DATE', *'\).*\('\)/\1$tdate\2/" $f
+      sed -i "s/^\(SetFormat( 'TIME', *'\).*\('\)/\1$ttime\2/" $f
+      sed -i "s/^\(SetFormat( 'YEAR', *'\).*\('\)/\1$tyear\2/" $f
+   done
+fi
+
+if [ "$@" = "update" ] ; then
+   cpfile
+   e_arrow2 "Updating Vundle"
+   gitclone https://github.com/VundleVim/Vundle.vim.git ~/.vim/bundle/Vundle.vim
+   e_arrow2 "Removing unused Plugins"
+   vim +"set viewoptions=" +PluginClean! +qall
+   e_arrow2 "Installing new Plugins"
+   vim +"set viewoptions=" +PluginInstall +qall
+   e_arrow2 "Updating Template Files according to $CONFFILE"
+   tfiles=( ~/.vim/bundle/bash-support.vim/bash-support/templates/Templates ~/.vim/bundle/awk-support.vim/awk-support/templates/Templates )
+   for f in ${tfiles[@]}; do 
+      sed -i "s/^\(SetMacro( 'AUTHOR', *'\).*\('\)/\1$tauthor\2/" $f
+      sed -i "s/^\(SetMacro( 'AUTHORREF', *'\).*\('\)/\1$tauthorref\2/" $f
+      sed -i "s/^\(SetMacro( 'COMPANY', *'\).*\('\)/\1$tcompany\2/" $f
+      sed -i "s/^\(SetMacro( 'COPYRIGHT', *'\).*\('\)/\1$tcopyright\2/" $f
+      sed -i "s/^\(SetMacro( 'EMAIL', *'\).*\('\)/\1$temail\2/" $f
+      sed -i "s/^\(SetMacro( 'LICENSE', *'\).*\('\)/\1$tlicense\2/" $f
+      sed -i "s/^\(SetMacro( 'ORGANIZATION', *'\).*\('\)/\1$torganization\2/" $f
+      sed -i "s/^\(SetFormat( 'DATE', *'\).*\('\)/\1$tdate\2/" $f
+      sed -i "s/^\(SetFormat( 'TIME', *'\).*\('\)/\1$ttime\2/" $f
+      sed -i "s/^\(SetFormat( 'YEAR', *'\).*\('\)/\1$tyear\2/" $f
    done
 fi
 }
 
 ptmux() {
 name="tmux"
-if [ "$@" = "dotfile" ] ; then
-   dotfile="tmux.conf"
-   dotdir="tmux"
+dotfile="tmux.conf"
+dotdir="tmux"
+if [ "$@" = "install" ] ; then
+   cpfile
+   e_arrow2 "clone tmux-plugins"
+   gitclone https://github.com/tmux-plugins/tpm ~/.tmux/plugins/tpm
+   e_arrow2 "Installing Plugins"
+   tmux start-server
+   tmux new-session -d
+   ~/.tmux/plugins/tpm/scripts/install_plugins.sh
+fi
+
+if [ "$@" = "update" ] ; then
+   cpfile
+   e_arrow2 "update tmux-plugins"
+   gitclone https://github.com/tmux-plugins/tpm ~/.tmux/plugins/tpm
+   e_arrow2 "Removing unused Plugins"
+   tmux start-server
+   tmux new-session -d
+   ~/.tmux/plugins/tpm/scripts/clean_plugins.sh
+   e_arrow2 "Updating Plugins"
+   ~/.tmux/plugins/tpm/scripts/update_plugin.sh
+fi
+
+}
+
+pshell() {
+name="shell"
+dotfile="shellrc"
+dotdir="shell"
+if [ "$@" = "install" ] ; then
+   rc=${SHELL#*/bin/}rc
+   src="source ~/.$dotfile"
+   cpfile
+   if grep -xq "^source ~/.shellrc" ~/.$rc
+   then
+      e_warn "$dotfile is already sourced in $rc. Adding nothing to $rc"
+   else
+      e_arrow2 "adding to $rc"
+      echo $src >> ~/.$rc
+   fi
+   mkdir -p ~/.$dotdir
+   e_arrow2 "downloading git-prompt/completion"
+   wget  -N -P ~/.$name/ https://raw.githubusercontent.com/git/git/master/contrib/completion/git-prompt.sh
+   wget  -N -P ~/.$name/ https://raw.githubusercontent.com/git/git/master/contrib/completion/git-completion.bash
+   wget  -N -O ~/.$name/_git https://raw.githubusercontent.com/git/git/master/contrib/completion/git-completion.zsh
+fi
+
+if [ "$@" = "update" ] ; then
+   rc=${SHELL#*/bin/}rc
+   src="source ~/.$dotfile"
+   cpfile
+   if grep -xq "^source ~/.shellrc" ~/.$rc
+   then
+      e_arrow2 "$dotfile is already sourced in $rc. Adding nothing to $rc"
+   else
+      e_arrow2 "adding to $rc"
+      echo $src >> ~/.$rc
+   fi
+   mkdir -p ~/.$dotdir
+   e_arrow2 "downloading git-prompt/completion"
+   wget -N -P ~/.$name/ https://raw.githubusercontent.com/git/git/master/contrib/completion/git-prompt.sh
+   wget -N -P ~/.$name/ https://raw.githubusercontent.com/git/git/master/contrib/completion/git-completion.bash
+   wget -N -O ~/.$name/_git https://raw.githubusercontent.com/git/git/master/contrib/completion/git-completion.zsh
 fi
 }
 
 psol() {
-name="solarized-colors"
-if [ "$@" = "dotfile" ] ; then
-   dotfile="solrc"
-   dotdir="sol"
+name="sol"
+dotfile="solrc"
+dotdir="sol"
+if [ "$@" = "install" ] ; then
+   cpfile
+   mkdir -p ~/.$dotdir
+   cp -r $name/colors ~/.sol/
+   e_arrow2 "clone dircolors"
+   gitclone https://github.com/seebi/dircolors-solarized ~/.sol/dircolors
+   e_arrow2 "Updating Template Files according to $CONFFILE"
+   sed -i "s/tsolarizecolor/$tsolarizecolor/g" ~/.$dotfile
+fi
+
+if [ "$@" = "update" ] ; then
+   cpfile
+   mkdir -p ~/.$dotdir
+   cp -r $name/colors ~/.sol/
+   e_arrow2 "update dircolors"
+   gitclone https://github.com/seebi/dircolors-solarized ~/.sol/dircolors
+   e_arrow2 "Updating Template Files according to $CONFFILE"
+   sed -i "s/tsolarizecolor/$tsolarizecolor/g" ~/.$dotfile
+fi
+}
+
+pgitlab() {
+name="gitlab"
+dotfile="python-gitlab.cfg"
+dotdir=""
+
+if [ "$@" = "install" ] ; then
+   cpfile
+   e_arrow2 "install python-gitlab"
+   pip install --user --upgrade python-gitlab
+   e_arrow2 "Updating Template Files according to $CONFFILE"
+   sed -i "s/skelplace/$tgitlabname/g" ~/.$dotfile
+   sed -i "s|skelurl|$tgitlaburl|g" ~/.$dotfile
+   sed -i "s/skeltoken/$tgitlabtoken/g" ~/.$dotfile
+fi
+
+if [ "$@" = "update" ] ; then
+   cpfile
+   e_arrow2 "install python-gitlab"
+   pip install --user --upgrade python-gitlab
+   e_arrow2 "Updating Template Files according to $CONFFILE"
+   sed -i "s/skelplace/$tgitlabname/g" ~/.$dotfile
+   sed -i "s|skelurl|$tgitlaburl|g" ~/.$dotfile
+   sed -i "s/skeltoken/$tgitlabtoken/g" ~/.$dotfile
 fi
 }
 
@@ -103,24 +231,28 @@ install() {
    check_config
    backup_dotfiles org
    remove_dotfiles
-   install_dotfiles
-   print_messages
+   do_dotfiles install
+   e_success "pdot successfully installed! Please login again."
 }
 
 update() {
    update_git
    check_config
    backup_dotfiles old
-   update_dotfiles
-   print_messages
+   do_dotfiles update
+   e_success "pdot successfully updated"
 }
 
 uninstall() {
    remove_dotfiles
-   unbackup_dotfiles_org
-   print_messages
+   unbackup_dotfiles
+   e_success "pdot successfully uninstalled"
 }
 
+update_git() {
+   e_header "Updating pdot repository"
+   git pull
+}
 check_config() {
    e_header "Checking if configure script was run..."
    if [ ! -e "$CONFFILE" ]; then
@@ -140,32 +272,75 @@ backup_dotfiles() {
    e_arrow "Backing up to $BACKUP"
    for i in "${!progs[@]}"; do
       ${progs[$i]} dotfile
-      cp $HOME/.$dotfile $BACKUP/
-      e_arrow2 "Backed up .$dotfile"
+      if [ -e "$HOME/.$dotfile" ] ; then
+         cp $HOME/.$dotfile $BACKUP/
+         e_success "Backed up .$dotfile"
+      fi
    done
    e_success "Backed up everything"
+}
+
+unbackup_dotfiles() {
+   e_header "Restoring dotfiles"
+   BACKUP="$BACKDIR-org/"
+   e_arrow "Restoring from $BACKUP"
+   for i in "${!progs[@]}"; do
+      ${progs[$i]} dotfile
+      if [ -e "$BACKUP/.$dotfile" ] ; then
+         cp $BACKUP/.$dotfile $HOME/
+         e_success "restored original .$dotfile"
+      fi
+   done
+   e_success "Restored everything"
 }
 
 remove_dotfiles() {
    e_header "Removing old dotfiles"
    for i in "${!progs[@]}"; do
       ${progs[$i]} dotfile
-      #rm -rf ~/.$dotdir
-      #rm ~/.$dotfile
-      e_arrow2 "removed ~/.$dotfile and ~/.$dotdir"
+      rm -f ~/.$dotfile
+      e_success "removed ~/.$dotfile"
+      if [ ! "$dotdir" = "" ] ; then
+	 rm -rf ~/.$dotdir
+         e_success "removed ~/.$dotdir"
+      fi
    done
    e_success "Removed all dotfiles"
 }
 
-install_dotfiles() {
-   e_header "Installing Dotfiles"
+do_dotfiles() {
+   e_header "$@ Dotfiles"
    for i in "${!progs[@]}"; do
-      ${progs[$i]}
-      e_arrow "installing $name"
-      ${progs[$i]} install
-      e_success "installed $name"
+      ${progs[$i]} name
+      e_arrow "$@ $name"
+      ${progs[$i]} $@
+      e_success "done with $name"
    done
-   e_success "Installed dotfiles"
+   e_success "finished $@"
 }
 
-(install)
+helpmessage() {
+   e_header "Usage"
+   echo "dotfiles [install|update|uninstall]"
+   exit 1
+}
+if [ $# -eq 0 ]; then
+    helpmessage
+fi
+mode=$1
+if [ $# -gt 1 ]; then
+   progs=()
+   for item in ${@:2}; do
+      progs+=("p$item")
+   done
+   echo ${progs[@]}
+fi
+if [[ $mode == "install" ]]; then
+  install
+elif [[ $mode == "update" ]]; then
+  update
+elif [[ $mode == "uninstall" ]]; then
+  uninstall
+else
+   helpmessage
+fi
